rename module for arguments processing
